Add resource file fetching in imon
On InstanceStatusUpdated if the source node has a avail-up and
started instance and we have a secondary instance, and the
remote file is more recent than ours, do fetch.

This patch does the fetch inline in the event handler.
A coming patch will make the fetch asynchronously to not block
the event loop.